requirements.txt
# Web framework
fastapi==0.110.2
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# AI / OpenAI
openai==1.12.0

# Resume parsing
python-docx==1.1.2
pdfplumber==0.11.0

# Template rendering
jinja2==3.1.4

# PDF generation (WeasyPrint + deps)
weasyprint==61.2
cairocffi==1.7.0
cffi==1.16.0

# Utilities
pydantic==2.7.1
typing-extensions==4.12.2